moves files to new architecture
